+++ src/main/java/io/dgraph/DgraphClient.java
@@ -38,30 +38,32 @@ import java.net.URL;
  * @author Abhimanyu Singh Gaur
  */
 public class DgraphClient {
+  private static final String gRPC_AUTHORIZATION_HEADER_NAME = "authorization";
 
   private final DgraphAsyncClient asyncClient;
 
   /**
-   * Implementation of a DgraphClientStub using grpc.
+   * Creates a gRPC stub that can be used to construct clients to connect with Slash GraphQL.
    *
-   * <p>Returns DgraphClientStub, that can be used to query/mutate Slash GraphQL endpoints
-   *
-   * <p>Takes the Slash GraphQL Endpoint and and apiKey as input.
+   * @param slashEndpoint The url of the Slash GraphQL endpoint. Example:
+   *                      https://your-slash-instance.cloud.dgraph.io/graphql
+   * @param apiKey        The API key used to connect to your Slash GraphQL instance.
+   * @return A new DgraphGrpc.DgraphStub object to be used with DgraphClient/DgraphAsyncClient.
    */
   public static DgraphGrpc.DgraphStub clientStubFromSlashEndpoint(
       String slashEndpoint, String apiKey) throws MalformedURLException {
-    URL url = new URL(slashEndpoint);
-    String[] parts = url.getHost().split("[.]", 2);
+    String[] parts = new URL(slashEndpoint).getHost().split("[.]", 2);
     if (parts.length < 2) {
       throw new MalformedURLException("Invalid Slash URL.");
     }
-    String grpcAddress = parts[0] + ".grpc." + parts[1];
+    String gRPCAddress = parts[0] + ".grpc." + parts[1];
 
     Metadata metadata = new Metadata();
-    metadata.put(Metadata.Key.of("authorization", Metadata.ASCII_STRING_MARSHALLER), apiKey);
+    metadata.put(Metadata.Key.of(gRPC_AUTHORIZATION_HEADER_NAME,
+        Metadata.ASCII_STRING_MARSHALLER), apiKey);
     return MetadataUtils.attachHeaders(
         DgraphGrpc.newStub(
-            ManagedChannelBuilder.forAddress(grpcAddress, 443).useTransportSecurity().build()),
+            ManagedChannelBuilder.forAddress(gRPCAddress, 443).useTransportSecurity().build()),
         metadata);
   }
 
@@ -71,7 +73,7 @@ public class DgraphClient {
    * <p>A single client is thread safe.
    *
    * @param stubs - an array of grpc stubs to be used by this client. The stubs to be used are
-   *     chosen at random per transaction.
+   *              chosen at random per transaction.
    */
   public DgraphClient(DgraphGrpc.DgraphStub... stubs) {
     this.asyncClient = new DgraphAsyncClient(stubs);
@@ -175,7 +177,7 @@ public class DgraphClient {
    * access JWT and a refresh JWT, which will be stored in the jwt field of this class, and used for
    * authorizing all subsequent requests sent to the server.
    *
-   * @param userid the id of the user who is trying to login, e.g. Alice
+   * @param userid   the id of the user who is trying to login, e.g. Alice
    * @param password the password of the user
    */
   public void login(String userid, String password) {
