+++ src/test/java/io/dgraph/DgraphClientTest.java
@@ -15,9 +15,6 @@
  */
 package io.dgraph;
 
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertTrue;
-
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.google.protobuf.ByteString;
@@ -25,10 +22,16 @@ import io.dgraph.DgraphProto.Mutation;
 import io.dgraph.DgraphProto.Operation;
 import io.dgraph.DgraphProto.Response;
 import io.dgraph.DgraphProto.TxnContext;
+
+import java.net.MalformedURLException;
 import java.util.Collections;
 import java.util.Map;
+
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+
+import static org.testng.Assert.*;
+import static org.testng.Assert.fail;
 
 /**
  * @author Edgar Rodriguez-Diaz
@@ -149,5 +152,24 @@ public class DgraphClientTest extends DgraphIntegrationTest {
     DgraphProto.Version v = dgraphClient.checkVersion();
     assertTrue(v.getTag().length() > 0);
     assertEquals(v.getTag().charAt(0), 'v');
+  }
+
+  @Test
+  public void testFromSlashEndpoint_ValidURL() {
+    try {
+      DgraphClient.clientStubFromSlashEndpoint("https://your-slash-instance.cloud.dgraph" +
+          ".io/graphql", "");
+    } catch (MalformedURLException e) {
+      fail(e.getMessage());
+    }
+  }
+
+  @Test
+  public void testFromSlashEndpoint_InValidURL() {
+    try {
+      DgraphClient.clientStubFromSlashEndpoint("https://a-bad-url", "");
+      fail("Invalid Slash URL should not be accepted.");
+    } catch (MalformedURLException e) {
+    }
   }
 }
